refactor: extract delegation enforcement tests from kernel.rs
Move the `delegation_enforcement` test module (8 tests, 220 lines)
from kernel.rs to tests/kernel_delegation.rs as an integration test.
This continues the kernel.rs simplification effort from PR #825.

kernel.rs: 4297 -> 4077 lines. Ratchet ceiling: 4300 -> 4080.
